+++ Najwieksze-skojarzenie-w-grafie-dwudizelnym/FileController.h
@@ -9,8 +9,8 @@
 
 namespace FileController
 {
-	Graph ConvertInputFile(std::string fileName, bool outputFile = false, int* sumOfWages = nullptr);
-	void Save(std::optional<Graph> graphToSave, bool hasBeenSolved, double sumOfWages, std::string fileName);
+	Graph ConvertInputFile(std::string fileName, bool outputFile = false, float* sumOfWages = nullptr);
+	void Save(std::optional<Graph> graphToSave, bool hasBeenSolved, float sumOfWages, std::string fileName);
 
     template<typename T>
     static T ReadOne(std::ifstream& file) {
